CLIENT-4391: Create index to use “integer” instead of “numeric” starting with Aerospike server >= 8.1.2 (#228)

+++ AerospikeDemo/QuerySum.cs
@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2023 Aerospike, Inc.
+ * Copyright 2012-2026 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -72,7 +72,7 @@ end
 
 			try
 			{
-				IndexTask task = client.CreateIndex(policy, args.ns, args.set, indexName, binName, IndexType.NUMERIC);
+				IndexTask task = client.CreateIndex(policy, args.ns, args.set, indexName, binName, IndexType.INTEGER);
 				task.Wait();
 			}
 			catch (AerospikeException ae)
